checkout----50%

+++ src/main/java/com/example/sv/Model/CartItem.java
@@ -7,7 +7,7 @@ public class CartItem {
     private long productId;
     private String name;
     private String productCategory;
-    private double price;
+    private int price;
     private int quantity;
     private Blob image;
 
@@ -43,11 +43,11 @@ public class CartItem {
         this.name = name;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
 
@@ -58,4 +58,7 @@ public class CartItem {
     public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
+
+
+
 }
